requirements.txt
discord.py==1.7.3
asyncio
pytz==2021.1